Friend request Changes
added new socket event for updating friend request.
added parameter to the proccessedFriendRequest so in the client will be able to
check wether the user is the sender or receiver of the request.
added new function for updating the friend request update,restore,delete or block.

also fixed issue where in addStatusToUser function cannot find status.